Move generator base class into its own module so distributed can pick it up
